started working on new screen class and fixed useing longs instead of ints DO NOT USE

+++ src/main/java/net/fractalinfinity/bettermaps/mapidlocations.java
@@ -16,9 +16,9 @@ import java.util.concurrent.ConcurrentHashMap;
 public class mapidlocations implements Listener {
 
     private static final Plugin plugin = Bettermaps.getPlugin(Bettermaps.class);
-    public static ConcurrentHashMap<Long, Location> maplocations = new ConcurrentHashMap<>();
+    public static ConcurrentHashMap<Integer, Location> maplocations = new ConcurrentHashMap<>();
 
-    public static void initialize(ConcurrentHashMap<Long, Location> maploc) {
+    public static void initialize(ConcurrentHashMap<Integer, Location> maploc) {
         if (maploc != null) maplocations = maploc;
         plugin.getServer().getPluginManager().registerEvents(new mapidlocations(), plugin);
     }
@@ -27,18 +27,18 @@ public class mapidlocations implements Listener {
     public void onHangingBreak(HangingBreakEvent event) {
         if (event.getEntity() instanceof ItemFrame || event.getEntity() instanceof GlowItemFrame && ((ItemFrame) event.getEntity()).getItem().getType() == Material.FILLED_MAP) {
             MapMeta meta  = ((MapMeta)((ItemFrame) event.getEntity()).getItem().getItemMeta());
-            if (meta != null) maplocations.remove((long) meta.getMapId());
+            if (meta != null) maplocations.remove((int) meta.getMapId());
         }
     }
     @EventHandler
     public void onItemframeinteract(PlayerItemFrameChangeEvent e){
         if (e.getAction() == PlayerItemFrameChangeEvent.ItemFrameChangeAction.PLACE && e.getItemStack().getType() == Material.FILLED_MAP){
             MapMeta mapmeta = ((MapMeta) e.getItemStack().getItemMeta());
-            maplocations.put((long) mapmeta.getMapId(),e.getItemFrame().getLocation());
+            maplocations.put(mapmeta.getMapId(),e.getItemFrame().getLocation());
         }
         else if (e.getAction() == PlayerItemFrameChangeEvent.ItemFrameChangeAction.REMOVE && e.getItemStack().getType() == Material.FILLED_MAP){
             MapMeta mapmeta = ((MapMeta) e.getItemStack().getItemMeta());
-            maplocations.remove((long) mapmeta.getMapId());
+            maplocations.remove(mapmeta.getMapId());
         }
 
     }
